Add source.js
The injection file, source.js has been added. The contents have to
be injected to the browser console of web.whatsapp.com. The send()
function takes two arguments: message and count. The function sends
the message a specified number of times, as determined by the value
of the count argument. The function can be invoked from console after
selecting the user in the web app.