Prevent freezing and enable early termination
Previously, the web app was prone to freezing due to the simultaneous
firing of all events, which caused excessive load. Moreover, it was not
possible to stop the process until the message queue was empty. To
address these issues, an interval has been introduced between each
event dispatch. This allows the termination of the firing process at
any point. A new function, stop(), has been implemented to facilitate
this functionality.